Add indexing lecture. Revert to static navbar.
Static navbar is nicer because if I just crank up the font size on full screen mode in Chrome, the side links get out of the way and I am left with just the full screen large view of the section.  In many regards this might be better than using slides, because I don't have to worry about stuff getting chopped off the edge or the bottom.